refactor cards controller
